move resources to api.py
* create parametres/api.py
* also move authorization and authentication classes to api.py
* add comment about signal to auto-create api keys
* update urls
* update _meta api to use django10, get_all_field_names is deprecated
* use [f.name for f in get_fields()] instead
